requirements.txt
astroid==2.4.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.4
idna==2.8
isort==5.6.4
jdatetime==3.6.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.19.4
pandas==1.1.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.22.0
selenium==3.141.0
six==1.15.0
ta==0.10.1
toml==0.10.2
urllib3==1.25.7
wrapt==1.12.1
xmltodict==0.13.0
